Add CLI config options
