+++ packages/css/cli/index.js
@@ -1,80 +1,55 @@
 #!/usr/bin/env node
-const fs = require("fs-extra");
-const path = require("path");
-const yamlToJson = require("yamljs");
-const jsonToSass = require("json-sass");
-const spawn = require("cross-spawn");
+"use strict";
+const meow = require("meow");
+const run = require("./run");
 
-function writeSassIndexByName(name, opts) {
-  spawn.sync(
-    "yarn",
-    [
-      "node-sass-chokidar",
-      `./dist/sass/${name}.scss`,
-      `./dist/${name}.css`,
-      "--source-map",
-      "true",
-      "--output-style",
-      "compact"
-    ],
-    opts
-  );
-  spawn.sync(
-    "yarn",
-    [
-      "node-sass-chokidar",
-      `./dist/sass/${name}.scss`,
-      `./dist/${name}.min.css`,
-      "--source-map",
-      "true",
-      "--output-style",
-      "compressed"
-    ],
-    opts
-  );
-  spawn.sync("gzip", ["-k", `./dist/${name}.css`], opts);
-  spawn.sync("gzip", ["-k", `./dist/${name}.min.css`], opts);
-}
+const cli = meow(
+  `
+    Usage
+      $ skeletor
+ 
+    Options
+      --config, -C     Specify config file/path
+      --output, -O     Specify output folder/path 
+      --gzip           Produce gzipped outputs (system command)
+      --no-sass        Disable production of Sass core
+      --no-minify      Disable production of minified outputs
+      --no-sourcemap   Disable production of sourcemaps
+ 
+    Examples
+      $ skeletor --config config.yml --output dist
+      Specify an alternate path to config file (default: ./skeletor.yml)
+`,
+  {
+    flags: {
+      config: {
+        type: "string",
+        alias: "C",
+        default: "skeletor.yml"
+      },
+      output: {
+        type: "string",
+        alias: "O",
+        default: "skeletor"
+      },
+      gzip: {
+        type: "boolean",
+        default: false
+      },
+      minify: {
+        type: "boolean",
+        default: true
+      },
+      sourcemap: {
+        type: "boolean",
+        default: true
+      },
+      sass: {
+        type: "boolean",
+        default: true
+      }
+    }
+  }
+);
 
-async function main({ config = "skeletor.yml" } = {}) {
-  const cwd = process.cwd();
-  const local = path.join(__dirname, "..");
-  const paths = {
-    config: path.join(cwd, config)
-  };
-  const asJsonObject = yamlToJson.load(paths.config);
-  const asJsonString = JSON.stringify(asJsonObject, undefined, 2);
-  const srcDir = path.join(local, "src");
-  const destDir = path.join(local, "dist");
-  const destSassDir = path.join(destDir, "sass");
-  const tempJson = path.join(destDir, "config.json");
-  const sassConfig = path.join(destSassDir, "_config.scss");
-  await fs.remove(destDir);
-  await fs.copy(srcDir, destSassDir);
-  await fs.writeFile(tempJson, asJsonString);
-
-  await new Promise(resolve => {
-    fs.createReadStream(tempJson)
-      .pipe(jsonToSass({ prefix: "$skeletor: " }))
-      .pipe(fs.createWriteStream(sassConfig))
-      .on("finish", resolve);
-  });
-
-  const opts = { stdio: "inherit", cwd: local };
-  writeSassIndexByName("skeletor", opts);
-  writeSassIndexByName("skeletor.grid", opts);
-  writeSassIndexByName("skeletor.utilities", opts);
-  writeSassIndexByName("skeletor.vars", opts);
-
-  await fs.remove(path.join(destDir, "sass/grid"));
-  await fs.remove(path.join(destDir, "sass/utilities"));
-  await fs.remove(path.join(destDir, "sass/skeletor.scss"));
-  await fs.remove(path.join(destDir, "sass/skeletor.grid.scss"));
-  await fs.remove(path.join(destDir, "sass/skeletor.utilities.scss"));
-  await fs.remove(path.join(destDir, "sass/skeletor.vars.scss"));
-
-  await fs.copy(destDir, path.join(cwd, "public/dist"));
-}
-
-// run
-main();
+run(cli.flags);
