+++ packages/css/package.json
@@ -42,6 +42,7 @@
     "cross-spawn": "^7.0.1",
     "fs-extra": "^8.1.0",
     "json-sass": "^1.3.5",
+    "meow": "^6.0.0",
     "yamljs": "^0.3.0"
   }
 }
